Add test xsds for reading
I've felt bad about the lack of tests I've had so far.

It has been a struggle, because I presume I shouldn't include things
from VMware, but they're the obvious things to write tests against.

The json files that I generate are kind of a test.

The xsds are complicated enough, and can be made generic that I think
it's worth bringing them in as test cases.