Parse if a field/type is required
It's quite verbose, but I'm getting into the swing of parsing the XML.

I'd prefer if it wasn't so verbose, but I'm happy just to get it
working.

I've looked into a few ways of parsing this.

XML crates that parse the xml directly into annotated rust structs would
be the easiest, but seem to lack the namespace support and possibly
don't handle these xml documents not following a predictable format all
the times.