Parse content-type
I've noticed there seem to be 2 types of annotations. I will see if I
need to split them out.

Type 1 is on a complex type and doesn't seem to have required,
deprecated or modifiable fields, but does have content-type.

Type 2 is on a field and has required, deprecated and modifiable fields,
but does not have content-type.

I think this is the last data I shall try and get out of the
annotations.