Hard code the different query parameters
The information isn't in the spec, so I had to hardcode them.

I think it's useful to be able to know when things are booleans, arrays
or integers.

I made fields string type

It has to be a comma separated list of fields (ie array of strings), but
the tooling wasn't cooperating.

Possibly `style: simple` would have worked
https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md#parameterObject,
but it isn't supported by the openapiv3 library I'm using.