Continue to ignore ovf schemas
This isn't a change of intended behaviour. The path they sit on changed,
and as such they weren't being ignored.

https://github.com/ccouzens/vcloud-rest-openapi/issues/4

I believe the reason I ignored them in the first place was they didn't
appear to be following the conventions of the rest of the schema, and as
I didn't have exposure to them it would be unclear if my mapping would
work.